Textures, animations.
Now textures can be just dropped on the folder. It has to be named:
(NAME)_(TYPE)(XXX).png (where (NAME) can have _ (underscores) as separators, (TYPE) is either image or anim and XXX is a 3 digit number starting at 001).

Four styles of Animation:
 - Do once.
 - Do once forward and once backwards.
 - Loop that starts at 0 when reaches frames.size()
 - Loop hat goes forward and backwards non-stop.

XML has heavily changed. Interactable HAS TO HAVE the same name both in config and in interactive_colliders_1 so they link up.

Speed has to be decided in config.xml.