Animations of objects without colliders done.
They have to be named anim_(NAME)_animXXX.png (where XXX starts at 001 and goes up for each frame in order).

They can be configured in xml (attributes speed and animstyle).

They also accept an static image, it's used when it's not playing the animation. It has to be named anim_(NAME)_static.png